Change LED color according to potentiometer volts
This program converts the analog input from ADC channel 13 (PTB3) using software trigger continuously. 
PTB3 is connect to the potentiometer POT1 at the TFC-Shield board. When sdjusting POT1, the voltage of PTB3 varies from 0 volts to 3.3 volts.The value of this voltage is used to control the tri-color LEDs. 

When the potentiometer is turned, the LEDs colour changes.